+++ locust/exception.py
@@ -4,6 +4,9 @@ class LocustError(Exception):
 class ResponseError(LocustError):
     pass
 
+class CatchResponseError(LocustError):
+    pass
+
 class InterruptLocust(Exception):
     """
     Exception that will interrupt a Locust when thrown inside a task