+++ locust/test/runtests.py
@@ -1,5 +1,5 @@
 import unittest
-from locust_class import TestLocustClass, TestSubLocust, TestWebLocustClass
+from locust_class import TestLocustClass, TestSubLocust, TestWebLocustClass, TestCatchResponse
 from test_stats import TestRequestStats, TestRequestStatsWithWebserver, TestInspectLocust
 
 if __name__ == '__main__':